+++ tests/test_distances.py
@@ -3,6 +3,7 @@ import tensorflow as tf
 
 from tensorflow_similarity.distances import CosineDistance
 from tensorflow_similarity.distances import EuclidianDistance
+from tensorflow_similarity.distances import ManhattanDistance
 
 
 def angular_distance_np(feature):
@@ -64,3 +65,45 @@ def test_euclidean_opposite():
     d = EuclidianDistance()
     vals = d(a)
     assert tf.reduce_all(tf.math.equal(vals, tf.constant([[1e-8, 2.0],[2.0, 1e-8]])))
+
+def test_manhattan():
+    a = tf.convert_to_tensor([
+        [0.0, 0.0],
+        [0.0, 1.0],
+        [1.0, 1.0],
+        [3.0, 0.0]
+    ])
+    d = ManhattanDistance()
+    vals = d(a)
+    expected = tf.constant([
+        [0.0, 1.0, 2.0, 3.0],
+        [1.0, 0.0, 1.0, 4.0],
+        [2.0, 1.0, 0.0, 3.0],
+        [3.0, 4.0, 3.0, 0.0]
+    ])
+    assert tf.reduce_all(tf.math.equal(vals, expected))
+
+def test_manhattan_axis_0():
+    a = tf.convert_to_tensor([
+        [0.0, 0.0],
+        [0.0, 1.0],
+        [1.0, 1.0],
+        [3.0, 0.0]
+    ])
+    d = ManhattanDistance()
+    vals = d(a, axis=0)
+    expected = tf.constant([[0.0, 4.0], [4.0, 0.0]])
+    assert tf.reduce_all(tf.math.equal(vals, expected))
+
+def test_manhattan_same():
+    a = tf.convert_to_tensor([[1.0, 1.0], [1.0, 1.0]])
+    d = ManhattanDistance()
+    vals = d(a)
+    assert tf.round(tf.reduce_sum(vals)) == 0
+
+
+def test_manhattan_opposite():
+    a = tf.convert_to_tensor([[0.0, 1.0], [0.0, -1.0]])
+    d = ManhattanDistance()
+    vals = d(a)
+    assert tf.reduce_all(tf.math.equal(vals, tf.constant([[0.0, 2.0],[2.0, 0.0]])))